Redo event handling system
The event handler is no longer supplied to the create function and is
instead passed on in every call to handle_events. The event handler can
now also be either a struct containing the function handle_event(Event)
or any function which takes which takes an event as the only parameter.
These functions can also return errors which will be correctly passed on
to the handle_events call.